Fix browser notifications
Capabilities actually need to be an array of strings. Also, browsers
require the actions capability before they'll deign to use the
notification daemon you've chosen for your system, so pop that in there
and pretend it's supported.

I also removed the intern stuff in favor of explicitly listing the
functions since that lets me jump to them.

Lastly, two of these signals are ones you send, not receive, so fix
that.